feat(mobile): feed summary component
